refactor: add NestedScrollView to HomeFragment to fix event not showing
This commit updates the HomeFragment to use a NestedScrollView to ensure that the event views are properly displayed. The previous implementation did not allow the user to scroll through the events, making it difficult to view all of the events.

BREAKING CHANGE: The HomeFragment layout has been updated to use a NestedScrollView. This may require changes to any custom views or layouts that are used in the HomeFragment.